separated string map storage
